bugfix: initialize Semaphore with Long

+++ modules/core/src/main/scala/turbolift/io/Semaphore.scala
@@ -28,6 +28,6 @@ sealed trait Semaphore:
 object Semaphore:
   private[turbolift] trait Unsealed extends Semaphore
 
-  def apply(initial: Int): Semaphore !! IO = create(initial)
-  def create(initial: Int): Semaphore !! IO = !!.impure(unsafeCreate(initial))
-  def unsafeCreate(initial: Int): Semaphore = new SemaphoreImpl(initial)
+  def apply(initial: Long): Semaphore !! IO = create(initial)
+  def create(initial: Long): Semaphore !! IO = !!.impure(unsafeCreate(initial))
+  def unsafeCreate(initial: Long): Semaphore = new SemaphoreImpl(initial)
